add finall commit

+++ src/main/java/by/anpoliakov/domainServices/PlayerRepository.java
@@ -2,12 +2,14 @@ package by.anpoliakov.domainServices;
 
 import by.anpoliakov.domain.Player;
 
+import java.math.BigDecimal;
+
 /** интерфейс с описанием действий над сущностью
  * @see Player */
 public interface PlayerRepository {
     void addPlayer(Player player);
     boolean existPlayerByLogin(String login);
-    boolean updateBalancePlayer(int player_id, Double newBalance);
+    boolean updateBalancePlayer(int player_id, BigDecimal newBalance);
     Player getPlayerByLoginAndPassword(String login, String password);
-    Double getBalancePlayer(int player_id);
+    BigDecimal getBalancePlayer(int player_id);
 }
